Async tasks dialog test
Re #194
